Add c-mode key binding for header guard
When in c mode (cc-mode), "C-c h" is added as a key binding to the
c-mode-map to insert a header guard like so.

	// my_file.h
	#ifndef __MY_FILE_H
	#define __MY_FILE_H

	#endif /* #ifndef __MY_FILE_H */

The function that does the insert is interactive, so it will prompt you
for the guard, offering the equivalent to the above for your file as the
default.